backend/requirements.txt
flask
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-dbapi
mysql-connector-python
